04_13_2009 Mac
git-svn-id: http://dim3.googlecode.com/svn/trunk@475 2c4a6428-2a2a-0410-ad13-7dde452bfe2e

+++ dim3Engine/Sources_View/effects_draw.c
@@ -40,7 +40,7 @@ extern server_type		server;
 extern view_type		view;
 
 extern bool effect_inview(effect_type *effect,int count);
-extern int distance_to_view_center(int x,int y,int z);
+extern double distance_to_view_center(int x,int y,int z);
 extern bool view_mesh_in_draw_list(int mesh_idx);
 
 /* =======================================================
